Feature/main fragment

+++ domain/src/main/java/prieto/fernando/domain/model/CompanyInfoDomainModel.kt.kt
@@ -6,5 +6,5 @@ data class CompanyInfoDomainModel(
     val founded: String,
     val employees: String,
     val launchSites: Int,
-    val valuation: Int
+    val valuation: Long
 )
